Maybe initialize the database
When you deploy your app for the first time, this script will do the
initial database migrations, and also (mostly) set up the auth extension
with email and password. Note that we are not configuring SMTP here
since that's currently impossible to do in a default way for every
deployment. In the future, we'll switch to webhooks so we can control it
from within the application itself.